refactor: Profile페이지이름에서 MyProfile로 이름변경
