+++ src/pages/MyProfile/components/MannerReview/index.tsx
