+++ src/stories/pages/MyProfile/PersonalInfo.stories.tsx
@@ -1,0 +1,15 @@
+import type { Meta, StoryObj } from '@storybook/react';
+
+import PersonalInfo from '@/pages/MyProfile/components/PersonalInfo';
+
+const meta: Meta<typeof PersonalInfo> = {
+  title: 'Pages/MyProfile/PersonalInfo',
+  tags: ['autodocs'],
+  component: PersonalInfo,
+};
+
+export default meta;
+
+type Story = StoryObj<typeof PersonalInfo>;
+
+export const Default: Story = {};
